+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
  - Layout in Flickable
  - LSP server with Live Preview and basic auto completion
  - The viewer tool gained the `--auto-reload` argument
+ - `Window.default-font-weight`
 
 ### Fixed
  - `Image::image-fit`'s `cover` and `contains` varient are fixed to match the CSS spec